+++ Debug/firmware/drivers/CoreI2C/i2c_interrupt.d
@@ -8,9 +8,9 @@ firmware/drivers/CoreI2C/i2c_interrupt.o: \
  ../firmware/drivers/CoreI2C/core_i2c.h \
  ../firmware/drivers/CoreI2C/../../hal/CortexM3/GNU/cpu_types.h \
  ../firmware/drivers/CoreI2C/../../CMSIS/m2sxxx.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cm3.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmInstr.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmFunc.h
+ ../firmware/drivers/CoreI2C/../../CMSIS/./core_cm3.h \
+ ../firmware/drivers/CoreI2C/../../CMSIS/././core_cmInstr.h \
+ ../firmware/drivers/CoreI2C/../../CMSIS/././core_cmFunc.h
 
 ../firmware/drivers/CoreI2C/../../hal/hal.h:
 
@@ -28,8 +28,8 @@ firmware/drivers/CoreI2C/i2c_interrupt.o: \
 
 ../firmware/drivers/CoreI2C/../../CMSIS/m2sxxx.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cm3.h:
+../firmware/drivers/CoreI2C/../../CMSIS/./core_cm3.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmInstr.h:
+../firmware/drivers/CoreI2C/../../CMSIS/././core_cmInstr.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmFunc.h:
+../firmware/drivers/CoreI2C/../../CMSIS/././core_cmFunc.h: