+++ Debug/firmware/drivers/mss_hpdma/mss_hpdma.d
@@ -3,9 +3,9 @@ firmware/drivers/mss_hpdma/mss_hpdma.o: \
  ../firmware/drivers/mss_hpdma/../../CMSIS/mss_assert.h \
  ../firmware/drivers/mss_hpdma/mss_hpdma.h \
  ../firmware/drivers/mss_hpdma/../../CMSIS/m2sxxx.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cm3.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmInstr.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmFunc.h
+ ../firmware/drivers/mss_hpdma/../../CMSIS/./core_cm3.h \
+ ../firmware/drivers/mss_hpdma/../../CMSIS/././core_cmInstr.h \
+ ../firmware/drivers/mss_hpdma/../../CMSIS/././core_cmFunc.h
 
 ../firmware/drivers/mss_hpdma/../../CMSIS/mss_assert.h:
 
@@ -13,8 +13,8 @@ firmware/drivers/mss_hpdma/mss_hpdma.o: \
 
 ../firmware/drivers/mss_hpdma/../../CMSIS/m2sxxx.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cm3.h:
+../firmware/drivers/mss_hpdma/../../CMSIS/./core_cm3.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmInstr.h:
+../firmware/drivers/mss_hpdma/../../CMSIS/././core_cmInstr.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmFunc.h:
+../firmware/drivers/mss_hpdma/../../CMSIS/././core_cmFunc.h: