+++ Debug/firmware/drivers/mss_nvm/mss_nvm.d
@@ -1,19 +1,19 @@
 firmware/drivers/mss_nvm/mss_nvm.o: ../firmware/drivers/mss_nvm/mss_nvm.c \
  ../firmware/drivers/mss_nvm/../../CMSIS/m2sxxx.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cm3.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmInstr.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmFunc.h \
+ ../firmware/drivers/mss_nvm/../../CMSIS/./core_cm3.h \
+ ../firmware/drivers/mss_nvm/../../CMSIS/././core_cmInstr.h \
+ ../firmware/drivers/mss_nvm/../../CMSIS/././core_cmFunc.h \
  ../firmware/drivers/mss_nvm/../../CMSIS/mss_assert.h \
  ../firmware/drivers/mss_nvm/../../CMSIS/system_m2sxxx.h \
  ../firmware/drivers/mss_nvm/mss_nvm.h
 
 ../firmware/drivers/mss_nvm/../../CMSIS/m2sxxx.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cm3.h:
+../firmware/drivers/mss_nvm/../../CMSIS/./core_cm3.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmInstr.h:
+../firmware/drivers/mss_nvm/../../CMSIS/././core_cmInstr.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmFunc.h:
+../firmware/drivers/mss_nvm/../../CMSIS/././core_cmFunc.h:
 
 ../firmware/drivers/mss_nvm/../../CMSIS/mss_assert.h: