+++ Debug/firmware/drivers/mss_sys_services/mss_sys_services.d
@@ -2,9 +2,9 @@ firmware/drivers/mss_sys_services/mss_sys_services.o: \
  ../firmware/drivers/mss_sys_services/mss_sys_services.c \
  ../firmware/drivers/mss_sys_services/mss_sys_services.h \
  ../firmware/drivers/mss_sys_services/../../CMSIS/m2sxxx.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cm3.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmInstr.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmFunc.h \
+ ../firmware/drivers/mss_sys_services/../../CMSIS/./core_cm3.h \
+ ../firmware/drivers/mss_sys_services/../../CMSIS/././core_cmInstr.h \
+ ../firmware/drivers/mss_sys_services/../../CMSIS/././core_cmFunc.h \
  ../firmware/drivers/mss_sys_services/mss_comblk_page_handler.h \
  ../firmware/drivers/mss_sys_services/mss_comblk.h \
  ../firmware/drivers/mss_sys_services/../../CMSIS/mss_assert.h
@@ -13,11 +13,11 @@ firmware/drivers/mss_sys_services/mss_sys_services.o: \
 
 ../firmware/drivers/mss_sys_services/../../CMSIS/m2sxxx.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cm3.h:
+../firmware/drivers/mss_sys_services/../../CMSIS/./core_cm3.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmInstr.h:
+../firmware/drivers/mss_sys_services/../../CMSIS/././core_cmInstr.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmFunc.h:
+../firmware/drivers/mss_sys_services/../../CMSIS/././core_cmFunc.h:
 
 ../firmware/drivers/mss_sys_services/mss_comblk_page_handler.h: