+++ Debug/src/main.d
@@ -2,9 +2,9 @@ src/main.o: ../src/main.c ../src/../inc/HM11/hm11.h \
  ../src/../inc/HM11/../../firmware/drivers/CoreUARTapb/core_uart_apb.h \
  ../src/../inc/HM11/../../firmware/drivers/CoreUARTapb/../../hal/CortexM3/GNU/cpu_types.h \
  ../src/../inc/HM11/../../firmware/CMSIS/m2sxxx.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cm3.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmInstr.h \
- C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmFunc.h \
+ ../src/../inc/HM11/../../firmware/CMSIS/./core_cm3.h \
+ ../src/../inc/HM11/../../firmware/CMSIS/././core_cmInstr.h \
+ ../src/../inc/HM11/../../firmware/CMSIS/././core_cmFunc.h \
  ../src/../inc/HM11/../CRC8/crc8.h ../src/../inc/MPU6050/mpu6050.h \
  ../src/../inc/MPU6050/../../firmware/CMSIS/m2sxxx.h \
  ../src/../inc/MPU6050/../../firmware/drivers/CoreI2C/core_i2c.h \
@@ -20,11 +20,11 @@ src/main.o: ../src/main.c ../src/../inc/HM11/hm11.h \
 
 ../src/../inc/HM11/../../firmware/CMSIS/m2sxxx.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cm3.h:
+../src/../inc/HM11/../../firmware/CMSIS/./core_cm3.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmInstr.h:
+../src/../inc/HM11/../../firmware/CMSIS/././core_cmInstr.h:
 
-C:/Microsemi/SoftConsole_v4.0/CMSIS/V4.3/Include/core_cmFunc.h:
+../src/../inc/HM11/../../firmware/CMSIS/././core_cmFunc.h:
 
 ../src/../inc/HM11/../CRC8/crc8.h: